refactor: group session-related utils into src/utils/session/
Move session.ts → session/params.ts, session-index.ts → session/derivation.ts,
session-store.ts → session/store.ts, session-resolver.ts → session/resolver.ts,
migration.ts → session/migration.ts, crypto.ts → session/crypto.ts. Add a
session/index.ts barrel so callers import a single module rather than six
sibling files.

No logic change; all 315 tests still pass and typecheck is clean.
